move tests to docker package, add necessary comment
